feat: refactor teacher management components; replace Main component with GET, implement ListTeachers, and enhance delete functionality with confirmation dialog
