+++ app/globals.css
@@ -118,6 +118,7 @@
     @apply border-border outline-ring/50;
   }
   body {
+    pointer-events: auto !important;
     @apply bg-background text-foreground;
   }
 }
