+++ components/admin/teachers/delete.tsx
@@ -1,42 +1,64 @@
 "use client";
+
+import { useState, cloneElement, ReactElement, ComponentProps } from "react";
 import BaseForm from "@/components/form/base-form";
 import {
-  AlertDialogFooter,
+  AlertDialog,
+  AlertDialogTrigger,
+  AlertDialogContent,
   AlertDialogHeader,
+  AlertDialogTitle,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogAction,
+  AlertDialogCancel,
 } from "@/components/ui/alert-dialog";
 import { Button } from "@/components/ui/button";
+import { Trash2 } from "lucide-react";
+
 import { deleteTeacher } from "@/lib/actions/teacher";
 import { useTeacherContext } from "@/lib/contexts/TeacherContext";
 import {
   DeleteTeacherSchema,
   teacherDeleteRenderFields,
 } from "@/lib/schemas/teachers";
-import {
-  AlertDialog,
-  AlertDialogAction,
-  AlertDialogCancel,
-  AlertDialogContent,
-  AlertDialogDescription,
-  AlertDialogTitle,
-  AlertDialogTrigger,
-} from "@/components/ui/alert-dialog";
-import { Trash2 } from "lucide-react";
-import { useState } from "react";
+
 interface DeleteProps {
   id: string;
+  children?: ReactElement<ComponentProps<typeof Button>>;
 }
 
-const Delete = ({ id }: DeleteProps) => {
+const Delete = ({ id, children }: DeleteProps) => {
   const [open, setOpen] = useState(false);
   const { deleteItem } = useTeacherContext();
+
+  const triggerButton = children ? (
+    cloneElement(children, {
+      onClick: (e) => {
+        setOpen(true);
+        if (children.props.onClick) {
+          children.props.onClick(e);
+        }
+      },
+    })
+  ) : (
+    <Button
+      variant="destructive"
+      size="sm"
+      onClick={() => setOpen(true)}
+      className="w-full"
+    >
+      <Trash2 className="h-4 w-4 mr-2" />
+      Delete
+    </Button>
+  );
+
   return (
     <AlertDialog open={open} onOpenChange={setOpen}>
-      <AlertDialogTrigger asChild>
-        <Button variant="destructive" size="sm">
-          <Trash2 className="h-4 w-4 mr-2" />
-          Delete
-        </Button>
+      <AlertDialogTrigger asChild onClick={(e) => e.stopPropagation()}>
+        {triggerButton}
       </AlertDialogTrigger>
+
       <AlertDialogContent>
         <AlertDialogHeader>
           <AlertDialogTitle>Are you absolutely sure?</AlertDialogTitle>
@@ -45,6 +67,7 @@ const Delete = ({ id }: DeleteProps) => {
             teacher account and remove their data from our servers.
           </AlertDialogDescription>
         </AlertDialogHeader>
+
         <AlertDialogFooter>
           <BaseForm
             initialState={{ success: false, errors: {} }}
@@ -52,6 +75,8 @@ const Delete = ({ id }: DeleteProps) => {
             action={deleteTeacher}
             schema={DeleteTeacherSchema}
             fields={teacherDeleteRenderFields}
+            defaultValues={{ id }}
+            actionType="delete"
             submitText={
               <AlertDialogAction className="bg-red-600 hover:bg-red-700">
                 Delete
@@ -67,9 +92,7 @@ const Delete = ({ id }: DeleteProps) => {
                 Cancel
               </AlertDialogCancel>
             }
-            actionType="delete"
             handleCancel={() => setOpen(false)}
-            defaultValues={{ id }}
             onSuccess={() => {
               deleteItem(id);
               setOpen(false);
