+++ components/admin/teachers/get.tsx
